Add readonly mode

+++ app/logical/maintenance.rb
@@ -2,6 +2,8 @@ module Maintenance
   module_function
 
   def daily
+    return if Danbooru.config.readonly_mode
+
     ignoring_exceptions { PostPruner.new.prune! }
     ignoring_exceptions { Upload.where('created_at < ?', 1.week.ago).delete_all }
     #ignoring_exceptions { ApiCacheGenerator.new.generate_tag_cache }
@@ -18,6 +20,8 @@ module Maintenance
   end
 
   def weekly
+    return if Danbooru.config.readonly_mode
+
     #ignoring_exceptions { ApproverPruner.prune! }
     #ignoring_exceptions { TagRelationshipRetirementService.find_and_retire! }
   end
